refacotr: clubMemberTier LeagueRecordEntity -> ClubMemberEntity로 이동
