+++ badminton-api/src/main/resources/data.sql
@@ -77,17 +77,17 @@ VALUES ('Badminton Club 1', 'This is the first badminton club.',
         '2024-01-10 09:00:00', '2024-01-10 09:00:00');
 
 -- ClubMember 데이터 삽입
-INSERT INTO club_member (clubId, memberId, role, deleted, banned, createdAt, modifiedAt)
-VALUES (1, 1, 'ROLE_OWNER', false, false, '2024-01-01 09:00:00', '2024-01-01 09:00:00'),
-       (1, 2, 'ROLE_MANAGER', false, false, '2024-01-02 09:00:00', '2024-01-02 09:00:00'),
-       (1, 3, 'ROLE_USER', false, false, '2024-01-03 09:00:00', '2024-01-03 09:00:00'),
-       (1, 4, 'ROLE_USER', false, false, '2024-01-04 09:00:00', '2024-01-04 09:00:00'),
-       (1, 5, 'ROLE_USER', false, false, '2024-01-05 09:00:00', '2024-01-05 09:00:00'),
-       (1, 6, 'ROLE_USER', false, false, '2024-01-06 09:00:00', '2024-01-06 09:00:00'),
-       (2, 7, 'ROLE_OWNER', false, false, '2024-01-07 09:00:00', '2024-01-07 09:00:00'),
-       (2, 8, 'ROLE_USER', false, false, '2024-01-08 09:00:00', '2024-01-08 09:00:00'),
-       (2, 9, 'ROLE_OWNER', false, false, '2024-01-09 09:00:00', '2024-01-09 09:00:00'),
-       (2, 10, 'ROLE_MANAGER', false, false, '2024-01-10 09:00:00', '2024-01-10 09:00:00');
+INSERT INTO club_member (clubId, memberId, role, tier, deleted, banned, createdAt, modifiedAt)
+VALUES (1, 1, 'ROLE_OWNER', 'GOLD', false, false, '2024-01-01 09:00:00', '2024-01-01 09:00:00'),
+       (1, 2, 'ROLE_MANAGER', 'SILVER',false, false, '2024-01-02 09:00:00', '2024-01-02 09:00:00'),
+       (1, 3, 'ROLE_USER', 'GOLD',false, false, '2024-01-03 09:00:00', '2024-01-03 09:00:00'),
+       (1, 4, 'ROLE_USER', 'BRONZE',false, false, '2024-01-04 09:00:00', '2024-01-04 09:00:00'),
+       (1, 5, 'ROLE_USER', 'SILVER',false, false, '2024-01-05 09:00:00', '2024-01-05 09:00:00'),
+       (1, 6, 'ROLE_USER', 'GOLD',false, false, '2024-01-06 09:00:00', '2024-01-06 09:00:00'),
+       (2, 7, 'ROLE_OWNER', 'BRONZE',false, false, '2024-01-07 09:00:00', '2024-01-07 09:00:00'),
+       (2, 8, 'ROLE_USER', 'GOLD',false, false, '2024-01-08 09:00:00', '2024-01-08 09:00:00'),
+       (2, 9, 'ROLE_OWNER', 'GOLD',false, false, '2024-01-09 09:00:00', '2024-01-09 09:00:00'),
+       (2, 10, 'ROLE_MANAGER', 'SILVER',false, false, '2024-01-10 09:00:00', '2024-01-10 09:00:00');
 
 
 -- League 데이터 삽입 (leagueLocation 필드 추가)
@@ -141,14 +141,14 @@ VALUES (1, 1, false, '2024-01-01 09:00:00', '2024-01-01 09:00:00', 1),
        (10, 5, false, '2024-01-20 09:00:00', '2024-01-20 09:00:00', 10);
 
 -- LeagueRecord 데이터 삽입 (티어 기준표 반영)
-INSERT INTO league_record (winCount, loseCount, drawCount, matchCount, tier, clubMemberId, createdAt, modifiedAt)
-VALUES (16, 3, 1, 20, 'GOLD', 1, '2024-01-01 09:00:00', '2024-01-01 09:00:00'),
-       (7, 2, 1, 10, 'SILVER', 2, '2024-01-02 09:00:00', '2024-01-02 09:00:00'),
-       (19, 1, 0, 20, 'GOLD', 3, '2024-01-03 09:00:00', '2024-01-03 09:00:00'),
-       (7, 3, 0, 10, 'SILVER', 4, '2024-01-04 09:00:00', '2024-01-04 09:00:00'),
-       (4, 5, 1, 10, 'BRONZE', 5, '2024-01-05 09:00:00', '2024-01-05 09:00:00'),
-       (17, 2, 1, 20, 'GOLD', 6, '2024-01-06 09:00:00', '2024-01-06 09:00:00'),
-       (3, 6, 1, 10, 'BRONZE', 7, '2024-01-07 09:00:00', '2024-01-07 09:00:00'),
-       (6, 3, 1, 10, 'SILVER', 8, '2024-01-08 09:00:00', '2024-01-08 09:00:00'),
-       (7, 2, 1, 10, 'SILVER', 9, '2024-01-09 09:00:00', '2024-01-09 09:00:00'),
-       (18, 1, 1, 20, 'GOLD', 10, '2024-01-10 09:00:00', '2024-01-10 09:00:00');
+INSERT INTO league_record (winCount, loseCount, drawCount, matchCount, clubMemberId, createdAt, modifiedAt)
+VALUES (16, 3, 1, 20,  1, '2024-01-01 09:00:00', '2024-01-01 09:00:00'),
+       (7, 2, 1, 10,  2, '2024-01-02 09:00:00', '2024-01-02 09:00:00'),
+       (19, 1, 0, 20, 3, '2024-01-03 09:00:00', '2024-01-03 09:00:00'),
+       (7, 3, 0, 10,  4, '2024-01-04 09:00:00', '2024-01-04 09:00:00'),
+       (4, 5, 1, 10,  5, '2024-01-05 09:00:00', '2024-01-05 09:00:00'),
+       (17, 2, 1, 20,  6, '2024-01-06 09:00:00', '2024-01-06 09:00:00'),
+       (3, 6, 1, 10,  7, '2024-01-07 09:00:00', '2024-01-07 09:00:00'),
+       (6, 3, 1, 10,  8, '2024-01-08 09:00:00', '2024-01-08 09:00:00'),
+       (7, 2, 1, 10, 9, '2024-01-09 09:00:00', '2024-01-09 09:00:00'),
+       (18, 1, 1, 20, 10, '2024-01-10 09:00:00', '2024-01-10 09:00:00');
